Fixes on question submission part 1 - still some errors
